requirements.txt
customtkinter
pystray
Pillow
